Paired_Exercises/inheritance-part1-exercises-pair/src/main/java/com/techelevator/SavingsAccount.java
package com.techelevator;
import java.math.BigDecimal;

public class SavingsAccount extends BankAccount{
	
	public SavingsAccount() {
	}
	
	@Override
	public BigDecimal withdraw(BigDecimal amountToWithdraw) {
		if (getBalance().compareTo(amountToWithdraw) == -1 ) {
			return getBalance();
		}
		else if (getBalance().subtract(amountToWithdraw).compareTo(new BigDecimal (150)) == -1) {	
			super.withdraw(new BigDecimal (2));
			super.withdraw(amountToWithdraw);
			return getBalance();
		}
		else {
			super.withdraw(amountToWithdraw);
			return getBalance();
		}
			
	}
}
